+++ backend/internal/interface/http/user/handler.go
@@ -1,7 +1,7 @@
 package user
 
 import (
-	"example.com/m/internal/infrastructure/oidc"
+	"example.com/m/internal/infrastructure/auth/oidc"
 	"example.com/m/internal/usecase/user"
 )
